fix: use Array.from() instead of spread syntax
This makes the production build and applications using this library functional again. The issue was likely in how Webpack (and other bundlers) would handle the `[...array]` notation during compiling. It resulted in an iterator object instead of an array which was exhausted too quickly. Hence, the React component could not render any fields.

Test this by building the example application and verify that all the `<input>` fields are filled.

Fixes #21