feat: add coverage limits for tests
Jest now calculates code coverage when running unit tests. Minimum coverage is 90 % which may be changed later if necessary. Some of the test cases were improved to satisfy the new coverage limits, hence the new snapshot.

Turned out that `className` prop for the <Form> component was typed as required which was not the case so it has been typed as optional now.